move bluetooth stuff to its own file
