+++ esp32-s3/main/CMakeLists.txt
@@ -1,3 +1,3 @@
-idf_component_register(SRCS "dankdryer.c" "hx711.c" "ota.c"
+idf_component_register(SRCS "ble.c" "dankdryer.c" "hx711.c" "ota.c"
                     PRIV_REQUIRES app_update bt driver esp_adc esp_app_format esp_driver_gpio esp_http_server esp_wifi json mqtt nvs_flash spi_flash
                     INCLUDE_DIRS "")
