+++ esp32-s3/main/ble.h
@@ -1,0 +1,10 @@
+#ifndef DANKDRYER_BLE
+#define DANKDRYER_BLE
+
+#define VERSION "0.0.9"
+#define DEVICE "dankdryer"
+#define CLIENTID DEVICE VERSION
+
+int setup_ble(void);
+
+#endif
